requirements.txt
Flask==3.1.2
gunicorn
Authlib
python-dotenv==1.2.1
psycopg2-binary
google-cloud-secret-manager==2.26.0
requests>=2.31.0
twilio>=9.0.0
google-generativeai>=0.8.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
